为网络请求类增加缓存类型

+++ DTKit/view/DTGridTableView/DTGridTableViewCell.m
@@ -11,8 +11,8 @@
 
 @interface DTGridTableViewCell()
 {
-    int _number;
-    int _height;
+    NSInteger _number;
+    NSInteger _height;
 }
 @end
